MOD-1011: Have added fix for FeadbackAllert component
