+++ components/FeedbackAlert/FeedbackAlert.tsx
@@ -31,32 +31,38 @@ export const FeedbackAlert: React.FC<FeedbackAlertPropTypes> = ({
   const contentColor = skin === 'darkGreen' || skin === 'black' ? 'C100' : 'A00';
 
   return (
-    <div className={cn(bgColor, 'flex items-center justify-center gap-3 p-3 sm:p-4 rounded-lg')}>
-      <DynamicIcon 
-        iconName={iconName}
-        stroke={getColorByName(contentColor)?.hex}
-        className="flex-shrink-0"
-        strokeWidth={1.4}
-      />
-      <div className='flex flex-col gap-1'>
-        {title && (
+    <div className={cn(bgColor, 
+      'flex flex-col gap-4 items-start justify-center',
+      'md:flex-row md:items-center sm:p-4 ',
+      'p-3 rounded-lg',
+    )}>
+      <div className='flex gap-3 items-center justify-center'>
+        <DynamicIcon 
+          iconName={iconName}
+          stroke={getColorByName(contentColor)?.hex}
+          className="flex-shrink-0"
+          strokeWidth={1.4}
+        />
+        <div className='flex flex-col gap-1'>
+          {title && (
+            <Typography 
+              size="lg" 
+              adaptive
+              weight='bold'
+              color={contentColor}
+            >
+              {title}
+            </Typography>
+          )}
+
           <Typography 
-            size="lg" 
+            size="sm" 
             adaptive
-            weight='bold'
             color={contentColor}
           >
-            {title}
+            {text}
           </Typography>
-        )}
-
-        <Typography 
-          size="sm" 
-          adaptive
-          color={contentColor}
-        >
-          {text}
-        </Typography>
+        </div>
       </div>
       {children}
     </div>
