--- src/stories/Components/FeedbackAlert/docs.ts
@@ -1,15 +1,0 @@
-export const argTypes = {
-  skin: {
-    control: 'select',
-    options: [
-      'green', 'yellow', 'red', 'grey', 'darkGreen', 'black',
-    ],
-    table: {
-      defaultValue: {
-        summary: 'green',
-      }
-    }
-  }
-}
-
-export const args = {};
